Help command supported on command handling and patch
Some commands can be used on a dm with bot. Mentioning the bot can also be used in place of a prefix. Help command now exported to command handling.

+++ src/commands/shutdown.js
@@ -5,15 +5,18 @@ module.exports = {
     permissionLevel: 3,
     parameters: {},
     execute(message, otherArguments){
+        if(!["text", "dm"].includes(message.channel.type)) return;
         switch(otherArguments.length){
             case 0:
             message.channel.send("Shutting down!").then(
                 sentMessage => {
                     sentMessage.client.destroy();
-            });
+                }
+            );
             break;
 
             default:
+            message.channel.send("Too many arguments supplied.");
             break;
         }
     }
